API: Объединение контроллеров для вопросов
